backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pillow==10.2.0
pyserial==3.5
python-dotenv==1.0.1
pydantic==2.6.3 